+++ data_fetching/JSONHandler.py
@@ -4,7 +4,7 @@ from datetime import date, datetime
 import json
 
 from WUBRG import COLOR_COMBINATIONS
-from utils import settings
+from utils.settings import DATA_DIR_LOC, DATA_DIR_NAME
 from utils.Logger import Logger
 from utils.Fetcher import Fetcher
 
@@ -49,7 +49,7 @@ class JSONHandler:
 
     def get_folder_path(self) -> str:
         """Returns the appropriate folder path, based on the properties of the object."""
-        path = os.path.join(settings.DATA_DIR_LOC, settings.DATA_DIR_NAME, self.SET, self.FORMAT)
+        path = os.path.join(DATA_DIR_LOC, DATA_DIR_NAME, self.SET, self.FORMAT)
         if self.DATE:
             return os.path.join(path, str(self.DATE))
         else:
