+++ data_fetching/RawDataFetcher.py
@@ -1,9 +1,9 @@
 from datetime import time, date, datetime, timedelta
 
 from utils.Logger import Logger
-from data_handling.utils.date_helper import get_prev_17lands_update_time, utc_today
+from data_fetching.utils.date_helper import get_prev_17lands_update_time, utc_today
 
-from data_handling.JSONHandler import JSONHandler
+from data_fetching.JSONHandler import JSONHandler
 from game_metadata.FormatMetadata import FormatMetadata
