+++ game_metadata/utils/consts.py
@@ -1,0 +1,7 @@
+# Card Consts
+RARITY_ALIASES: dict[str, str] = {
+    'common': "C",
+    'uncommon': "U",
+    'rare': "R",
+    'mythic': "M"
+}
